Added test data loading. Refactored id type in entities.

+++ app/src/main/java/by/dashkevichpavel/osteopath/model/Disfunction.kt
@@ -5,10 +5,10 @@ import by.dashkevichpavel.osteopath.persistence.DbContract
 import by.dashkevichpavel.osteopath.persistence.entity.DisfunctionEntity
 
 data class Disfunction(
-    val id: Int = 0,
+    val id: Long = 0,
     var description: String = "",
     var disfunctionStatusId: Int = DisfunctionStatus.WORK.id,
-    var customerId: Int = 0
+    var customerId: Long = 0
 ) {
     constructor(disfunctionEntity: DisfunctionEntity) : this(
         id = disfunctionEntity.id,
